implemented verification code validation
